opea-comps/services/megaservice/requirements.txt
fastapi>=0.100.0
uvicorn==0.34.0
requests>=2.31.0
pydantic>=2.0.0
python-multipart==0.0.20
opea-comps>=1.2.0